refactor: actions name
