+++ src/state/budgets/index.ts
@@ -19,8 +19,8 @@ export type Budgets = {
 };
 
 export { default as budgetsSelectors } from "./selectors";
-export { default as actions } from "./actions";
-export { default as types } from "./types";
+export { default as budgetsActions } from "./actions";
+export { default as budgetsTypes } from "./types";
 export { sampleState };
 
 export default reducer;
